feat: Improve data display, add switch for intensive optimizations

+++ friendica-find-missing-servers.sh
@@ -8,9 +8,11 @@ elif [[ -n $(type mysql) ]]; then
 else
 	exit
 fi
+intense_optimizations=${1:-"0"}
 db="friendica"
 tmpfile="/tmp/sitesdown.txt"
 idsdownfile="/tmp/idsdown.txt"
+usrfile="/tmp/usersdown.txt"
 url=friendica.example.net
 avatarfolder=/var/www/friendica/avatar
 avatarfolderescaped=${avatarfolder////\\/}
@@ -44,9 +46,9 @@ loop_2() {
 
 loop_3() {
 	baseurltrimmed=$(echo "${baseurl}" | sed -e "s/http[s]*:\/\///g")
-	printf "%s - %s@%s " "${lineb}" "${nick}" "${baseurltrimmed}"
 	#Find the pictures in the avatar folders and delete them
-	"${dbengine}" "${db}" -N -B -q -e "select \`photo\`, \`thumb\`, \`micro\` from \`contact\` where \`id\` = ${lineb}" | while read -r photo thumb micro; do
+	picturecount=0
+	while read -r photo thumb micro; do
 		#If stored in avatar folder
 		if grep -v -q "${url}/avatar" <(echo "${photo}"); then
 			#if [[ -z "${isavatar}" ]]
@@ -56,36 +58,73 @@ loop_3() {
 			rm -rfv "${thumbtrimmed}"
 			microtrimmed=$(echo "${micro}" | sed -e "s/https:\/\/${url}\/avatar/${avatarfolderescaped}/g" -e "s/\?ts.*//g")
 			rm -rfv "${microtrimmed}"
+			picturecount=1
 		fi
+	done < <("${dbengine}" "${db}" -N -B -q -e "select \`photo\`, \`thumb\`, \`micro\` from \`contact\` where \`id\` = ${id}")
+	postthreadcount=$("${dbengine}" "${db}" -N -B -q -e "create temporary table tmp_post_thread (select \`uri-id\` from \`post-thread\` where \`owner-id\` = ${id} or \`author-id\` = ${id} or \`causer-id\` = ${id}); delete h.* from \`post-thread\` h inner join \`tmp_post_thread\` t where h.\`uri-id\` = t.\`uri-id\`; select_row_count();" || echo 0)
+	postthreadusercount=$("${dbengine}" "${db}" -N -B -q -e "create temporary table tmp_post_thread_user (select \`uri-id\` from \`post-thread-user\` where \`owner-id\` = ${id} or \`author-id\` = ${id} or \`causer-id\` = ${id}); delete r.* from \`post-thread-user\` r inner join \`tmp_post_thread_user\` t where r.\`uri-id\` = t.\`uri-id\`; select_row_count();" || echo 0)
+	postusercount=$("${dbengine}" "${db}" -N -B -q -e "create temporary table tmp_post_user (select \`id\` from \`post-user\` where \`owner-id\` = ${id} or \`author-id\` = ${id} or \`causer-id\` = ${id}); delete u.* from \`post-user\` u inner join \`tmp_post_user\` t where u.\`id\` = t.\`id\`; select_row_count();" || echo 0)
+	posttagcount=$("${dbengine}" "${db}" -N -B -q -e "delete from \`post-tag\` where cid = ${id}; select_row_count();" || echo 0)
+	postcontentcount=$("${dbengine}" "${db}" -N -B -q -e "create temporary table tmp_post (select \`uri-id\` from \`post\` where \`owner-id\` = ${id} or \`author-id\` = ${id} or \`causer-id\` = ${id}); delete p.* from \`post-content\` p inner join \`tmp_post\` t where p.\`uri-id\` = t.\`uri-id\`; select_row_count();" || echo 0)
+	postcount=$("${dbengine}" "${db}" -N -B -q -e "create temporary table tmp_post (select \`uri-id\` from \`post\` where \`owner-id\` = ${id} or \`author-id\` = ${id} or \`causer-id\` = ${id}); delete p.* from \`post\` p inner join \`tmp_post\` t where p.\`uri-id\` = t.\`uri-id\`; select_row_count();" || echo 0)
+	photocount=$("${dbengine}" "${db}" -N -B -q -e "delete from \`photo\` where \`contact-id\` = ${id}; select_row_count();" || echo 0)
+	contactcount=$("${dbengine}" "${db}" -N -B -q -e "delete from \`contact\` where \`id\` = ${id}; select_row_count();" || echo 0)
+	apcontactcount=$("${dbengine}" "${db}" -N -B -q -e "delete from \`apcontact\` where \`uri-id\` = ${id}; select_row_count();" || echo 0)
+	diasporacontactcount=$("${dbengine}" "${db}" -N -B -q -e "delete from \`diaspora-contact\` where \`uri-id\` = ${id}; select_row_count();" || echo 0)
+	while read -r tmp_picturecount tmp_postthreadcount tmp_postthreadusercount tmp_postusercount tmp_posttagcount tmp_postcontentcount tmp_postcount tmp_photocount tmp_contactcount tmp_apcontactcount tmp_diasporacontactcount; do
+		picturecount=$((picturecount + tmp_picturecount))
+		postthreadcount=$((postthreadcount + tmp_postthreadcount))
+		postthreadusercount=$((postthreadusercount + tmp_postthreadusercount))
+		postusercount=$((postusercount + tmp_postusercount))
+		posttagcount=$((posttagcount + tmp_posttagcount))
+		postcontentcount=$((postcontentcount + tmp_postcontentcount))
+		postcount=$((postcount + tmp_postcount))
+		photocount=$((photocount + tmp_photocount))
+		contactcount=$((contactcount + tmp_contactcount))
+		apcontactcount=$((apcontactcount + tmp_apcontactcount))
+		diasporacontactcount=$((diasporacontactcount + tmp_diasporacontactcount))
+	done <"${usrfile}"
+	response_left=$(printf "%s %s@%s " "${id}" "${nick}" "${baseurltrimmed}")
+	response=$(printf "%spicture:%s " "${response}" "${picturecount}")
+	response=$(printf "%spost-thread:%s " "${response}" "${postthreadcount}")
+	response=$(printf "%spost-thread-user:%s " "${response}" "${postthreadusercount}")
+	response=$(printf "%spost-user:%s " "${response}" "${postusercount}")
+	response=$(printf "%spost-tag:%s " "${response}" "${posttagcount}")
+	response=$(printf "%spost-content:%s " "${response}" "${postcontentcount}")
+	response=$(printf "%spost:%s " "${response}" "${postcount}")
+	response=$(printf "%sphoto:%s " "${response}" "${photocount}")
+	response=$(printf "%scontact:%s " "${response}" "${contactcount}")
+	response=$(printf "%sapcontact:%s " "${response}" "${apcontactcount}")
+	response=$(printf "%sdiaspora-contact:%s " "${response}" "${diasporacontactcount}")
+	echo "${picturecount}" "${postthreadcount} ${postthreadusercount} ${postusercount} ${posttagcount} ${postcontentcount} ${postcount} ${photocount} ${contactcount} ${apcontactcount} ${diasporacontactcount}" >"${usrfile}"
+	#Previous line clearance
+	#Measure length of string, blank only the excess
+	final_string_length_left="${#response_left}"
+	final_string_length_right="${#response}"
+	final_string_length=$((final_string_length_left + final_string_length_right))
+	blank_string=""
+	columns_length="${COLUMNS}"
+	while [[ "${final_string_length}" -gt "${columns_length}" ]]; do
+		columns_length=$((columns_length + COLUMNS))
 	done
-	postthreadcount=$("${dbengine}" "${db}" -N -B -q -e "create temporary table tmp_post_thread (select \`uri-id\` from \`post-thread\` where \`owner-id\` = ${lineb} or \`author-id\` = ${lineb} or \`causer-id\` = ${lineb}); delete h.* from \`post-thread\` h inner join \`tmp_post_thread\` t where h.\`uri-id\` = t.\`uri-id\`; select row_count();")
-	postthreadusercount=$("${dbengine}" "${db}" -N -B -q -e "create temporary table tmp_post_thread_user (select \`uri-id\` from \`post-thread-user\` where \`owner-id\` = ${lineb} or \`author-id\` = ${lineb} or \`causer-id\` = ${lineb}); delete r.* from \`post-thread-user\` r inner join \`tmp_post_thread_user\` t where r.\`uri-id\` = t.\`uri-id\`; select row_count();")
-	postusercount=$("${dbengine}" "${db}" -N -B -q -e "create temporary table tmp_post_user (select \`id\` from \`post-user\` where \`owner-id\` = ${lineb} or \`author-id\` = ${lineb} or \`causer-id\` = ${lineb}); delete u.* from \`post-user\` u inner join \`tmp_post_user\` t where u.\`id\` = t.\`id\`; select row_count();")
-	posttagcount=$("${dbengine}" "${db}" -N -B -q -e "delete from \`post-tag\` where cid = ${lineb}; select row_count();")
-	postcontentcount=$("${dbengine}" "${db}" -N -B -q -e "create temporary table tmp_post (select \`uri-id\` from \`post\` where \`owner-id\` = ${lineb} or \`author-id\` = ${lineb} or \`causer-id\` = ${lineb}); delete p.* from \`post-content\` p inner join \`tmp_post\` t where p.\`uri-id\` = t.\`uri-id\`; select row_count();")
-	postcount=$("${dbengine}" "${db}" -N -B -q -e "create temporary table tmp_post (select \`uri-id\` from \`post\` where \`owner-id\` = ${lineb} or \`author-id\` = ${lineb} or \`causer-id\` = ${lineb}); delete p.* from \`post\` p inner join \`tmp_post\` t where p.\`uri-id\` = t.\`uri-id\`; select row_count();")
-	photocount=$("${dbengine}" "${db}" -N -B -q -e "delete from \`photo\` where \`contact-id\` = ${lineb}; select row_count();")
-	contactcount=$("${dbengine}" "${db}" -N -B -q -e "delete from \`contact\` where \`id\` = ${lineb}; select row_count();")
-	apcontactcount=$("${dbengine}" "${db}" -N -B -q -e "delete from \`apcontact\` where \`uri-id\` = ${lineb}; select row_count();")
-	diasporacontactcount=$("${dbengine}" "${db}" -N -B -q -e "delete from \`diaspora-contact\` where \`uri-id\` = ${lineb}; select row_count();")
-	printf "post-thread: %s " "${postthreadcount}"
-	printf "post-thread-user: %s " "${postthreadusercount}"
-	printf "post-user: %s " "${postusercount}"
-	printf "post-tag: %s " "${posttagcount}"
-	printf "post-content: %s " "${postcontentcount}"
-	printf "post: %s " "${postcount}"
-	printf "photo: %s " "${photocount}"
-	printf "contact: %s " "${contactcount}"
-	printf "apcontact: %s " "${apcontactcount}"
-	printf "diaspora-contact: %s " "${diasporacontactcount}"
-	printf "\r\n"
+	blank_string_length=$((columns_length - final_string_length))
+	for ((count = 0; count < "${blank_string_length}"; count++)); do
+		blank_string=$(printf "%s " "${blank_string}")
+	done
+	response=$(printf "%s%s%s" "${response_left}" "${blank_string}" "${response}")
+	for ((count = 0; count < "${final_string_length}"; count++)); do
+		response=$(printf "%s\b" "${response}")
+	done
+	printf "%s\r" "${response}"
 }
 
 #Check if our dependencies are installed
 if [[ -n $(type curl) && -n "${dbengine}" && -n $(type "${dbengine}") && -n $(type date) ]]; then
 	date
-	"${dbengine}" "${db}" -N -B -q -e "alter table \`contact\` add index if not exists \`contact_baseurl\` (\`baseurl\`)"
-	"${dbengine}" "${db}" -N -B -q -e "alter table \`post-user\` add index if not exists \`post_user_id\` (\`author-id\`, \`causer-id\`, \`owner-id\`)"
+	if [[ "${intense_optimizations}" -gt 0 ]]; then
+		"${dbengine}" "${db}" -N -B -q -e "alter table \`contact\` add index if not exists \`contact_baseurl\` (\`baseurl\`)"
+		"${dbengine}" "${db}" -N -B -q -e "alter table \`post-user\` add index if not exists \`post_user_id\` (\`author-id\`, \`causer-id\`, \`owner-id\`)"
+	fi
 	if [[ ! -f "${tmpfile}" ]]; then
 		echo "Listing sites"
 		siteslist=$("${dbengine}" "${db}" -N -B -q -e "select distinct baseurl, protocol from contact where baseurl != ''" | sort -b -f -n | sed -e "s/http:/https:/g" | uniq -i)
@@ -117,24 +156,30 @@ if [[ -n $(type curl) && -n "${dbengine}" && -n $(type "${dbengine}") && -n $(ty
 		u=$(sort -n "${idsdownfile}" | uniq)
 		echo "${u}" >"${idsdownfile}"
 	fi
-	while read -r lineb nick baseurl; do
+	touch "${usrfile}"
+	echo "0 0 0 0 0 0 0 0 0 0 0" >"${usrfile}"
+	while read -r id nick baseurl; do
 		#The community no longer exists, delete
-		loop_3 "${lineb}" "${nick}" "${baseurl}" &
+		loop_3 "${id}" "${nick}" "${baseurl}" &
 		if [[ $(jobs -r -p | wc -l) -ge $(($(getconf _NPROCESSORS_ONLN) / 2)) ]]; then
 			wait -n
 		fi
 		wait
 	done <"${idsdownfile}"
+	printf "\r\n"
 	rm "${tmpfile}" 2>/dev/null
 	rm "${idsdownfile}" 2>/dev/null
-	"${dbengine}" "${db}" -N -B -q -e "alter table \`post-thread\` auto_increment = 1"
-	"${dbengine}" "${db}" -N -B -q -e "alter table \`post-thread-user\` auto_increment = 1"
-	"${dbengine}" "${db}" -N -B -q -e "alter table \`post-user\` auto_increment = 1"
-	"${dbengine}" "${db}" -N -B -q -e "alter table \`post-tag\` auto_increment = 1"
-	"${dbengine}" "${db}" -N -B -q -e "alter table \`post\` auto_increment = 1"
-	"${dbengine}" "${db}" -N -B -q -e "alter table \`photo\` auto_increment = 1"
-	"${dbengine}" "${db}" -N -B -q -e "alter table \`contact\` auto_increment = 1"
-	"${dbengine}" "${db}" -N -B -q -e "alter table \`contact\` drop index \`contact_baseurl\`"
-	"${dbengine}" "${db}" -N -B -q -e "alter table \`post-user\` drop index \`post_user_id\`"
+	rm "${usrfile}" 2>/dev/null
+	"${dbengine}" "${db}" -N -B -q -e "alter table \`post-thread\` auto_increment = 1; \
+		alter table \`post-thread-user\` auto_increment = 1; \
+		alter table \`post-user\` auto_increment = 1; \
+		alter table \`post-tag\` auto_increment = 1; \
+		alter table \`post\` auto_increment = 1; \
+		alter table \`photo\` auto_increment = 1; \
+		alter table \`contact\` auto_increment = 1"
+	if [[ "${intense_optimizations}" -gt 0 ]]; then
+		"${dbengine}" "${db}" -N -B -q -e "alter table \`contact\` drop index \`contact_baseurl\`"
+		"${dbengine}" "${db}" -N -B -q -e "alter table \`post-user\` drop index \`post_user_id\`"
+	fi
 	date
 fi
